feat: app router

+++ lib/core/constants/breakpoints.dart
@@ -1,5 +1,5 @@
 class Breakpoints {
   static const int mobile = 768;
-  static const double tablet = 900;
-  static const double desktop = 1200;
+  static const int tablet = 900;
+  static const int desktop = 1200;
 }
